*** empty log message ***
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@511 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ src/io/impl/easyAVI.h
@@ -106,7 +106,7 @@ void easyAVI::openAVI(char *filename, int framerate, int xr, int yr)
 	AVIFileInit();
 
 #ifdef UNICODE
-	const size_t bufferLength = MultiByteToWideChar( CP_ACP, 0, filename, -1, 0, 0);
+	const int bufferLength = MultiByteToWideChar( CP_ACP, 0, filename, -1, 0, 0);
 	std::vector<WCHAR> buffer(bufferLength);
 	MultiByteToWideChar( CP_ACP, 0, filename, -1, &buffer[0], bufferLength);
 	LPCWSTR szFile = &buffer[0];
